Fix iOS 5 crash + refactor.

+++ Functions.h
@@ -8,35 +8,33 @@ BOOL CamRotateisOn;
 BOOL CamRotateLock;
 BOOL SyncOrientation;
 
-int rotationStyle;
-int orientationValue;
+NSInteger rotationStyle;
+NSInteger orientationValue;
 
-HaveCallback()
-{
-	GetPrefs()
-	GetBool(CamRotateisOn, @"CamRotateEnabled", YES)
-	GetBool(CamRotateLock, @"CamRotateLock", NO)
-	GetBool(SyncOrientation, @"SyncOrientation", NO)
-	GetInt(rotationStyle, @"RotationStyle", 2)
-	GetInt(orientationValue, @"OrientationValue", 1)
+HaveCallback() {
+    GetPrefs()
+    GetBool(CamRotateisOn, @"CamRotateEnabled", YES)
+    GetBool(CamRotateLock, @"CamRotateLock", NO)
+    GetBool(SyncOrientation, @"SyncOrientation", NO)
+    GetInt(rotationStyle, @"RotationStyle", 2)
+    GetInt(orientationValue, @"OrientationValue", 1)
 }
 
-int glyphOrientationOverride(int orig)
-{
-	if (CamRotateLock)
-		return orientationValue;
-	if (SyncOrientation) {
-		UIInterfaceOrientation orient = [UIDevice.currentDevice orientation];
-		switch (orient) {
-			case UIInterfaceOrientationPortrait:
-				return 1;
-			case UIInterfaceOrientationPortraitUpsideDown:
-				return 2;
-			case UIInterfaceOrientationLandscapeLeft:
-				return 4;
-			case UIInterfaceOrientationLandscapeRight:
-				return 3;
-		}
-	}
-	return orig;
+NSInteger glyphOrientationOverride(NSInteger orig) {
+    if (CamRotateLock)
+        return orientationValue;
+    if (SyncOrientation) {
+        UIInterfaceOrientation orient = [UIDevice.currentDevice orientation];
+        switch (orient) {
+            case UIInterfaceOrientationPortrait:
+                return 1;
+            case UIInterfaceOrientationPortraitUpsideDown:
+                return 2;
+            case UIInterfaceOrientationLandscapeLeft:
+                return 4;
+            case UIInterfaceOrientationLandscapeRight:
+                return 3;
+        }
+    }
+    return orig;
 }
